chore(winding): restructure platform-specific utils
