+++ packages/winding/x11/mod.ts
@@ -1,6 +1,6 @@
-import type { Library, LoadLibrary, UIEvent, Window } from "./src/../types.ts";
-import { x11functions, XEventMask, XEventType } from "./x11_ffi.ts";
-import { utf8Bytes, utf8CString as cString } from "./text_encoding.ts";
+import type { Library, LoadLibrary, UIEvent, Window } from "../types.ts";
+import { utf8Bytes, utf8CString as cString } from "../text_encoding.ts";
+import { x11functions, XEventMask, XEventType } from "./ffi.ts";
 
 // XStoreName sets the legacy WM_NAME property, which is read as Latin-1 by clients that don't
 // understand the EWMH _NET_WM_NAME/UTF8_STRING property set below. Encoding it as UTF-8 there
